# apollo-openapi.yaml
openapi: 3.0.1
info:
  title: Apollo OpenAPI
  description: |
    <p>Apollo配置中心OpenAPI接口文档</p>

    <h2>认证方式</h2>
    <p>所有 API 接口都需要通过 <code>Authorization</code> header 进行身份验证。</p>

    <h3>获取 Token 的方式：</h3>
    <ol>
      <li><b>Portal 管理界面获取</b>：登录 Portal → 管理员工具 → 开放平台授权管理 → 创建第三方应用，获取 Token。</li>
      <li><b>Token 格式</b>：<code>Authorization: token_value</code></li>
      <li><b>Token 权限</b>：按应用/环境/命名空间授予，建议不同用途分别创建。</li>
    </ol>

    <h3>使用示例</h3>
    <pre><code class="language-bash">curl -X GET "http://localhost:8070/openapi/v1/apps" \
    -H "Authorization: your_token_here"</code></pre>

  version: 1.0.0
security:
  - ApiKeyAuth: []
...
paths:
...
  /openapi/v1/apps/{appId}/envs/{env}/accesskeys:
    post:
      summary: 创建AccessKey
      operationId: createAccessKey
      deprecated: false
      description: ''
      tags:
        - AccessKey Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: operator
          in: query
          description: ''
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAccessKeyDTO'
                description: com.ctrip.framework.apollo.common.dto.AccessKeyDTO
              example:
                dataChangeCreatedBy: ''
                dataChangeLastModifiedBy: ''
                dataChangeCreatedByDisplayName: ''
                dataChangeLastModifiedByDisplayName: ''
                dataChangeCreatedTime: ''
                dataChangeLastModifiedTime: ''
                id: 0
                secret: ''
                appId: ''
                mode: 0
                enabled: false
          headers: {}
    get:
      summary: 查找AccessKey
      operationId: findAccessKeys
      deprecated: false
      description: ''
      tags:
        - AccessKey Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenAccessKeyDTO'
                  description: com.ctrip.framework.apollo.common.dto.AccessKeyDTO
              example:
                - dataChangeCreatedBy: ''
                  dataChangeLastModifiedBy: ''
                  dataChangeCreatedByDisplayName: ''
                  dataChangeLastModifiedByDisplayName: ''
                  dataChangeCreatedTime: ''
                  dataChangeLastModifiedTime: ''
                  id: 0
                  secret: ''
                  appId: ''
                  mode: 0
                  enabled: false
          headers: {}
  /openapi/v1/apps/{appId}/envs/{env}/accesskeys/{accessKeyId}:
    delete:
      summary: 删除AccessKey
      operationId: deleteAccessKey
      deprecated: false
      description: ''
      tags:
        - AccessKey Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: accessKeyId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
        - name: operator
          in: query
          description: ''
          required: false
          schema:
            type: string
      responses:
        '200':
          description: ''
  /openapi/v1/apps/{appId}/envs/{env}/accesskeys/{accessKeyId}/activation:
    put:
      summary: 启用AccessKey
      operationId: enableAccessKey
      deprecated: false
      description: ''
      tags:
        - AccessKey Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: accessKeyId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
        - name: mode
          in: query
          description: ''
          required: false
          example: 0
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: ''
  /openapi/v1/apps/{appId}/envs/{env}/accesskeys/{accessKeyId}/deactivation:
    put:
      summary: 禁用AccessKey
      operationId: disableAccessKey
      deprecated: false
      description: ''
      tags:
        - AccessKey Management
      parameters:
        - name: appId
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: env
          in: path
          description: ''
          required: true
          schema:
            type: string
        - name: accessKeyId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: ''
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Apollo OpenAPI访问令牌
        **获取方式：**
        1. 登录Apollo Portal管理界面
        2. 进入"管理员工具" -> "开放平台授权管理"
        3. 创建第三方应用，获取Token
        **使用格式：**
        ```
        Authorization: your_token_here
        ```
        **权限说明：**
        - Token具有对应应用的读写权限
        - 不同Token可能有不同的环境和命名空间权限
        - 建议为不同用途创建不同的Token
  schemas:
...
    OpenAccessKeyDTO:
      type: object
      properties:
        dataChangeCreatedBy:
          type: string
          description: ''
        dataChangeLastModifiedBy:
          type: string
          description: ''
        dataChangeCreatedByDisplayName:
          type: string
          description: ''
        dataChangeLastModifiedByDisplayName:
          type: string
          description: ''
        dataChangeCreatedTime:
          type: string
          description: ''
        dataChangeLastModifiedTime:
          type: string
          description: ''
        id:
          type: integer
          description: ''
          format: int64
        secret:
          type: string
          description: ''
        appId:
          type: string
          description: ''
        mode:
          type: integer
          description: ''
        enabled:
          type: boolean
          description: ''
...
servers: []