# spring-boot2/src/main/java/com/apollo/openapi/server/api/AccessKeyManagementApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.apollo.openapi.server.api;

import com.apollo.openapi.server.model.OpenAccessKeyDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "AccessKey Management", description = "the AccessKey Management API")
public interface AccessKeyManagementApi {

    default AccessKeyManagementApiDelegate getDelegate() {
        return new AccessKeyManagementApiDelegate() {};
    }

    /**
     * POST /openapi/v1/apps/{appId}/envs/{env}/accesskeys : 创建AccessKey
     *
     *
     * @param appId  (required)
     * @param env  (required)
     * @param operator  (optional)
     * @return  (status code 200)
     */
    @Operation(
        operationId = "createAccessKey",
        summary = "创建AccessKey",
        description = "",
        tags = { "AccessKey Management" },
        responses = {
            @ApiResponse(responseCode = "200", description = "", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OpenAccessKeyDTO.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/openapi/v1/apps/{appId}/envs/{env}/accesskeys",
        produces = { "application/json" }
    )
    default ResponseEntity<OpenAccessKeyDTO> createAccessKey(
        @Parameter(name = "appId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("appId") String appId,
        @Parameter(name = "env", description = "", required = true, in = ParameterIn.PATH) @PathVariable("env") String env,
        @Parameter(name = "operator", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "operator", required = false) String operator
    ) {
        return getDelegate().createAccessKey(appId, env, operator);
    }


    /**
     * DELETE /openapi/v1/apps/{appId}/envs/{env}/accesskeys/{accessKeyId} : 删除AccessKey
     *
     *
     * @param appId  (required)
     * @param env  (required)
     * @param accessKeyId  (required)
     * @param operator  (optional)
     * @return  (status code 200)
     */
    @Operation(
        operationId = "deleteAccessKey",
        summary = "删除AccessKey",
        description = "",
        tags = { "AccessKey Management" },
        responses = {
            @ApiResponse(responseCode = "200", description = "")
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/openapi/v1/apps/{appId}/envs/{env}/accesskeys/{accessKeyId}"
    )
    default ResponseEntity<Void> deleteAccessKey(
        @Parameter(name = "appId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("appId") String appId,
        @Parameter(name = "env", description = "", required = true, in = ParameterIn.PATH) @PathVariable("env") String env,
        @Parameter(name = "accessKeyId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accessKeyId") Long accessKeyId,
        @Parameter(name = "operator", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "operator", required = false) String operator
    ) {
        return getDelegate().deleteAccessKey(appId, env, accessKeyId, operator);
    }


    /**
     * PUT /openapi/v1/apps/{appId}/envs/{env}/accesskeys/{accessKeyId}/deactivation : 禁用AccessKey
     *
     *
     * @param appId  (required)
     * @param env  (required)
     * @param accessKeyId  (required)
     * @return  (status code 200)
     */
    @Operation(
        operationId = "disableAccessKey",
        summary = "禁用AccessKey",
        description = "",
        tags = { "AccessKey Management" },
        responses = {
            @ApiResponse(responseCode = "200", description = "")
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/openapi/v1/apps/{appId}/envs/{env}/accesskeys/{accessKeyId}/deactivation"
    )
    default ResponseEntity<Void> disableAccessKey(
        @Parameter(name = "appId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("appId") String appId,
        @Parameter(name = "env", description = "", required = true, in = ParameterIn.PATH) @PathVariable("env") String env,
        @Parameter(name = "accessKeyId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accessKeyId") Long accessKeyId
    ) {
        return getDelegate().disableAccessKey(appId, env, accessKeyId);
    }


    /**
     * PUT /openapi/v1/apps/{appId}/envs/{env}/accesskeys/{accessKeyId}/activation : 启用AccessKey
     *
     *
     * @param appId  (required)
     * @param env  (required)
     * @param accessKeyId  (required)
     * @param mode  (optional, default to 0)
     * @return  (status code 200)
     */
    @Operation(
        operationId = "enableAccessKey",
        summary = "启用AccessKey",
        description = "",
        tags = { "AccessKey Management" },
        responses = {
            @ApiResponse(responseCode = "200", description = "")
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/openapi/v1/apps/{appId}/envs/{env}/accesskeys/{accessKeyId}/activation"
    )
    default ResponseEntity<Void> enableAccessKey(
        @Parameter(name = "appId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("appId") String appId,
        @Parameter(name = "env", description = "", required = true, in = ParameterIn.PATH) @PathVariable("env") String env,
        @Parameter(name = "accessKeyId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("accessKeyId") Long accessKeyId,
        @Parameter(name = "mode", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "mode", required = false, defaultValue = "0") Integer mode
    ) {
        return getDelegate().enableAccessKey(appId, env, accessKeyId, mode);
    }


    /**
     * GET /openapi/v1/apps/{appId}/envs/{env}/accesskeys : 查找AccessKey
     *
     *
     * @param appId  (required)
     * @param env  (required)
     * @return  (status code 200)
     */
    @Operation(
        operationId = "findAccessKeys",
        summary = "查找AccessKey",
        description = "",
        tags = { "AccessKey Management" },
        responses = {
            @ApiResponse(responseCode = "200", description = "", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = OpenAccessKeyDTO.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/openapi/v1/apps/{appId}/envs/{env}/accesskeys",
        produces = { "application/json" }
    )
    default ResponseEntity<List<OpenAccessKeyDTO>> findAccessKeys(
        @Parameter(name = "appId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("appId") String appId,
        @Parameter(name = "env", description = "", required = true, in = ParameterIn.PATH) @PathVariable("env") String env
    ) {
        return getDelegate().findAccessKeys(appId, env);
    }

}